core: De-duplicate enums between Avm1Button and Avm2Button
Needed to appease clippy.
