+++ core/src/display_object.rs
@@ -40,8 +40,8 @@ pub use crate::display_object::container::{
     DisplayObjectContainer, Lists, TDisplayObjectContainer,
 };
 use crate::events::{ClipEvent, ClipEventResult};
-pub use avm1_button::Avm1Button;
-pub use avm2_button::{Avm2Button, ButtonTracking};
+pub use avm1_button::{Avm1Button, ButtonState, ButtonTracking};
+pub use avm2_button::Avm2Button;
 pub use bitmap::Bitmap;
 pub use edit_text::{AutoSizeMode, EditText, TextSelection};
 pub use graphic::Graphic;
